requirements.txt
# requirements file

google-ai-generativelanguage==0.2.0
google-api-core==2.11.0
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.11.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-generativeai==0.1.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
kfp==2.4.0
vertexai==0.0.1
python-dotenv==1.0.0
scikit-learn==1.2.2
mplcursors==0.5.2
ipympl==0.9.3
tqdm==4.65.0
protobuf==3.19.6
ipython
pandas==2.2.0
pyarrow==12.0.1
transformers==4.16.2
sentence-transformers==2.2.0
nltk==3.2.5
torch==2.1.2
rouge==1.0.1
numpy
sequence-evaluate==0.0.3
openai==1.10.0
tiktoken==0.5.2
pinecone-client==3.0.2
openpyxl==3.1.2
spacy
plotly==5.18
dash==2.15.0
dash-bootstrap-components==1.5.0
dash-tools==1.12.0
markdown==3.6
beautifulsoup4==4.12.3